third_party/WebKit/LayoutTests/inspector/console/console-message-from-inline-with-url-expected.txt

CONSOLE MESSAGE: line 8: foo
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token }
Tests that log message and syntax errors from inline scripts with sourceURL are logged into console, contains correct link and doesn't cause browser crash.

foo.js:4 foo
boo.js:1 Uncaught SyntaxError: Unexpected token }
    at addInlineWithSyntaxError (foo.js:12)
addInlineWithSyntaxError @ foo.js:12
